requirements.txt
asgiref-3.7.2 
django-5.0.2 
sqlparse-0.4.4 
typing-extensions-4.9.0
Pillow-10.2.0
sorl-thumbnail-12.3